refactor: controller test 패키지 수정
